# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyAdminPanel",
    "description": "X-UI style proxy management system with BFF architecture",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/rules": {
      "get": {
        "tags": [
          "Rules"
        ],
        "summary": "Get All Rules",
        "description": "Get all routing rules.",
        "operationId": "get_all_rules_api_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  },
                  "type": "array",
                  "title": "Response Get All Rules Api Rules Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Rules"
        ],
        "summary": "Create Rule",
        "description": "Create new routing rule.",
        "operationId": "create_rule_api_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/rules/{rule_id}": {
      "get": {
        "tags": [
          "Rules"
        ],
        "summary": "Get Rule",
        "description": "Get rule by ID.",
        "operationId": "get_rule_api_rules__rule_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Rules"
        ],
        "summary": "Update Rule",
        "description": "Update existing rule.",
        "operationId": "update_rule_api_rules__rule_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rules"
        ],
        "summary": "Delete Rule",
        "description": "Delete rule.",
        "operationId": "delete_rule_api_rules__rule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/dashboard": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get Dashboard Stats",
        "description": "Get dashboard statistics.\nReturns CPU, memory, bandwidth, user counts, and system info.",
        "operationId": "get_dashboard_stats_api_system_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/system/backup": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Backup Database",
        "description": "Download database backup.\nReturns the SQLite database file for download.",
        "operationId": "backup_database_api_system_backup_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/system/sync-traffic": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Sync Traffic",
        "description": "Manually trigger traffic sync from Core Service.\nUpdates traffic statistics for all users.",
        "operationId": "sync_traffic_api_system_sync_traffic_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/system/check-expired": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Check Expired Users",
        "description": "Manually check and disable expired users.",
        "operationId": "check_expired_users_api_system_check_expired_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/system/admin/profile": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get Admin Profile",
        "description": "Get current admin profile.",
        "operationId": "get_admin_profile_api_system_admin_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Update Admin Profile",
        "description": "Update admin profile (username, password, avatar).",
        "operationId": "update_admin_profile_api_system_admin_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/settings/": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get System Settings",
        "description": "Get current system settings.",
        "operationId": "get_system_settings_api_settings__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update System Settings",
        "description": "Update system settings.",
        "operationId": "update_system_settings_api_settings__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/settings/generate-credentials": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Generate Test Credentials",
        "description": "Test endpoint to generate username/password based on current patterns.\nFor shadowsocks (ss), returns an encryption method as username.\nFor socks5/http, generates username based on pattern.\n\nQuery parameter:\n- protocol: \"socks5\", \"http\", or \"ss\" (default: \"socks5\")",
        "operationId": "generate_test_credentials_api_settings_generate_credentials_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "socks5",
              "title": "Protocol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminResponse": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "username",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "AdminResponse"
      },
      "AdminUpdate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 6
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "title": "AdminUpdate"
      },
...
      "DashboardStats": {
        "properties": {
          "cpu_usage": {
            "type": "number",
            "title": "Cpu Usage"
          },
          "memory_usage": {
            "type": "number",
            "title": "Memory Usage"
          },
          "total_memory": {
            "type": "integer",
            "title": "Total Memory"
          },
          "used_memory": {
            "type": "integer",
            "title": "Used Memory"
          },
          "bandwidth_up": {
            "type": "integer",
            "title": "Bandwidth Up"
          },
          "bandwidth_down": {
            "type": "integer",
            "title": "Bandwidth Down"
          },
          "online_users": {
            "type": "integer",
            "title": "Online Users"
          },
          "total_users": {
            "type": "integer",
            "title": "Total Users"
          },
          "system_version": {
            "type": "string",
            "title": "System Version"
          },
          "uptime": {
            "type": "string",
            "title": "Uptime"
          }
        },
        "type": "object",
        "required": [
          "cpu_usage",
          "memory_usage",
          "total_memory",
          "used_memory",
          "bandwidth_up",
          "bandwidth_down",
          "online_users",
          "total_users",
          "system_version",
          "uptime"
        ],
        "title": "DashboardStats"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "type": "object",
        "required": [
          "name",
          "content"
        ],
        "title": "RuleCreate"
      },
      "RuleResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "content",
          "id",
          "priority",
          "created_at",
          "updated_at"
        ],
        "title": "RuleResponse"
      },
      "RuleUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "type": "object",
        "title": "RuleUpdate"
      },
      "SuccessResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "SuccessResponse"
      },
      "SystemSettingsResponse": {
        "properties": {
          "default_protocol": {
            "type": "string",
            "title": "Default Protocol",
            "default": "socks5"
          },
          "default_expiration_days": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Default Expiration Days",
            "default": 30
          },
          "default_max_send_byte": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Max Send Byte",
            "default": 0
          },
          "default_max_receive_byte": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Max Receive Byte",
            "default": 0
          },
          "default_send_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Send Limit",
            "default": 0
          },
          "default_receive_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Receive Limit",
            "default": 0
          },
          "default_max_conn_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Max Conn Count",
            "default": 0
          },
          "username_pattern": {
            "type": "string",
            "title": "Username Pattern",
            "default": "LLL###"
          },
          "password_pattern": {
            "type": "string",
            "title": "Password Pattern",
            "default": "LLL###"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "SystemSettingsResponse"
      },
      "SystemSettingsUpdate": {
        "properties": {
          "default_protocol": {
            "type": "string",
            "title": "Default Protocol",
            "default": "socks5"
          },
          "default_expiration_days": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Default Expiration Days",
            "default": 30
          },
          "default_max_send_byte": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Max Send Byte",
            "default": 0
          },
          "default_max_receive_byte": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Max Receive Byte",
            "default": 0
          },
          "default_send_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Send Limit",
            "default": 0
          },
          "default_receive_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Receive Limit",
            "default": 0
          },
          "default_max_conn_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Default Max Conn Count",
            "default": 0
          },
          "username_pattern": {
            "type": "string",
            "title": "Username Pattern",
            "default": "LLL###"
          },
          "password_pattern": {
            "type": "string",
            "title": "Password Pattern",
            "default": "LLL###"
          }
        },
        "type": "object",
        "title": "SystemSettingsUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}